% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanEpiRisk.R
\docType{package}
\name{CanEpiRisk}
\alias{CanEpiRisk-package}
\alias{CanEpiRisk}
\title{CanEpiRisk}
\description{
Lifetime Risk Calculation
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kyojifurukawa.github.io/CanEpiRisk/}
}

}
