% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CER.R
\name{CER}
\alias{CER}
\title{Calculate the cumulative excess risk due to radiation exposure}
\usage{
CER(exposure, reference, riskmodel, option)
}
\arguments{
\item{exposure}{a list object that specifies the exposure scenario, which contains 'agex' (a single value or a vector for age(s) at exposure), 'doseGy' (a single value or a vector of dose(s) in Gy), and 'sex' (1 or 2 for male or female).}

\item{reference}{a list object that specifies the baseline information of the reference population, which contains data.frame objects named 'baseline' for baseline rates of the target endpoint and 'mortality' for all cause mortality rates.}

\item{riskmodel}{a list object that specifies the risk model, which contains two list objects named 'err' for excess relative rate model and 'ear' for excess absolute rate model, each of which contains a vector 'para' for model parameter estimates, a matrix 'var' for the variance covariance matrix, and a function 'f' to compute the excess risk given a parameter vector and exposure information (e.g., dose, age at exposure, sex, attained age).}

\item{option}{a list object that specifies optional settings for risk calculation, which contains an integer value 'maxage' for the maximum age to follow up, a value 'err_wgt' for the weight for risk transfer (1=err, 0=ear), an integer value 'n_mcsamp' for the number of Monte Carlo samples, and alpha for the significance level (default=0.05).}
}
\value{
information of calculated risk (vector)
}
\description{
Calculate the cumulative excess risk due to radiation exposure
}
\examples{
   # The following examples use default data provided in the CanEpiRisk package
   # allsolid mortality, Region-1, male, 0.1Gy at age 5, LSS linear ERR

   # Cumulated excess risk for male exposed to 0.1 Gy at age 10
   #  followed up to age 90 with err transfer
   CER( agex=10, doseGy=0.1, sex=1, maxage=90 )

   # Cumulated excess risk for female exposed to 0.01 Gy/year at ages 10-19
   #  followed up to age 100 with 7:3 weights for ERR and EAR transfers
   CER( agex=10:19, doseGy=rep(0.01,10), sex=2, maxage=100, wgt=c(.7,.3) )

}
