% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Incidence}
\alias{Incidence}
\title{Cancer incidence rates of WHO global regions}
\format{
A list object of the five WHO global regions:
\describe{
\item{\verb{Aus-NZ Europe Northern America}}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
\item{\verb{Northern Africa - Western Asia}}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
\item{\verb{Latin America and Caribbean}}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
\item{\verb{Asia excl. Western Asia}}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
\item{\verb{Sub-Saharan Africa}}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
}
}
\source{
IARC
}
\usage{
Incidence
}
\description{
Cancer incidence rates of WHO global regions.
}
\details{
The list object for each region contains the 16 site-specific cancer mortality rates (\code{esophagus}, \code{stomach}, \code{colon}, \code{liver}, \code{pancreas}, \code{lung}, \code{breast}, \code{prostate}, \code{bladder}, \code{brainCNS}, \code{thyroid}, \code{all_leukaemia}, \code{all_cancer}, \code{allsolid-NMSC}, \code{allsolid}, \code{leukaemia}).

Each site-specific data.frame contains variables \code{age}, \code{male} and \code{female}.
}
\examples{

 names(Incidence)       # WHO global regions
 names(Incidence[[1]])  # Sites for which baseline incidence rates are available

 # Example 1: All solid cancer incidence rates for Region-1
 Incidence[[1]]$allsolid

 # Example 2: Leukaemia incidence rates for Region-3
 Incidence[[3]]$leukaemia

 # Example 3: A;;ll;-cause incidence rates for Region-5
 Incidence[[5]]$allcause

 # Example 4: plotting lung cancer incidence rates
 plot_refdata( dat=Incidence, outcome="lung", leg_pos=c(0.27,0.95) )

}
\seealso{
\link{Mortality}, \link{plot_refdata}
}
\keyword{datasets}
