% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{LSS_incidence}
\alias{LSS_incidence}
\title{LSS incidence risk models}
\format{
A list object of :
\describe{
  \item{allsolid}{a list object which contains risk model information (see Details)}
  \item{leukaemia}{a list object which contains risk model information (see Details)}
  \item{esophagus}{a list object which contains risk model information (see Details)}
  \item{stomach}{a list object which contains risk model information (see Details)}
  \item{colon}{a list object which contains risk model information (see Details)}
  \item{liver}{a list object which contains risk model information (see Details)}
  \item{lung}{a list object which contains risk model information (see Details)}
  \item{prostate}{a list object which contains risk model information (see Details)}
  \item{pancreas}{a list object which contains risk model information (see Details)}
  \item{bladder}{a list object which contains risk model information (see Details)}
  \item{breast}{a list object which contains risk model information (see Details)}
  \item{thyroid}{a list object which contains risk model information (see Details)}
  \item{brainCNS}{a list object which contains risk model information (see Details)}
}
}
\source{
IARC
}
\usage{
LSS_incidence
}
\description{
LSS incidence risk models for 13 cancer sites.
}
\details{
The list object for each risk model contains the 13 site-specific cancer incidence risk models derived from Life Span Study ("leukaemia", "esophagus", "stomach", "colon", "liver", "lung", "prostate", "pancreas", "bladder", "breast", "thyroid", "brainCNS").

Each site-specific a list object contains information for the risk model (a vector of parameter estimates "para", a matrix of variance-covariance matrix of parameter estimates "var" and a function to calculate the risk "f").
}
\examples{
 names(LSS_incidence)   # Sites for which LSS incidence risk models are available
 names(LSS_incidence$allsolid)    # Available dose response models
 LSS_incidence$allsolid$L$err     # Linear ERR model for all solid cancer incidence
 LSS_incidence$allsolid$L$ear     # Linear EAR model for all solid cancer incidence

 LSS_incidence$leukaemia$LQ$err   # Linear-quadratic ERR model for leukaemia incidence
 LSS_incidence$lung$L$err         # Linear ERR model for thyroid cancer incidence

 # Plotting LSS all solid cancer mortality risk model
 plot_riskmodel( rm=LSS_incidence$allsolid$L, title="LSS all solid cancer mortality, Linear",  leg_pos=c(0.4, 0.95) )

 # Plotting LSS Leukaemia incidence risk model
 plot_riskmodel( rm=LSS_incidence$leukaemia$LQ, title="LSS leukaemia incidence", ymax=c(1.5, .3), add=c(0.01,0) )

 # Plotting LSS thyroid cancer incidence risk model
 plot_riskmodel( rm=LSS_incidence$thyroid$L, title="LSS thyroid cancer incidence", ymax=c(.5, .3) )

}
\references{
Ozasa, K., Y. Shimizu, A. Suyama et al. Studies of the mortality of atomic bomb survivors, Report 14, 1950-2003: an overview of cancer and noncancer diseases. Radiat Res 177(3): 229-243 (2012).
}
\seealso{
[LSS_mortality], [plot_riskmodel()]
}
\keyword{datasets}
