% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Mortality}
\alias{Mortality}
\title{Cancer mortality rates of WHO global regions}
\format{
A list object of the five WHO global regions:
\describe{
  \item{`Aus-NZ Europe Northern America`}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
  \item{`Northern Africa - Western Asia`}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
  \item{`Latin America and Caribbean`}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
  \item{`Asia excl. Western Asia`}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
  \item{`Sub-Saharan Africa`}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
}
}
\source{
IARC
}
\usage{
Mortality
}
\description{
Cancer mortality rates of WHO global regions.
}
\details{
The list object for each region contains the 18 site-specific cancer mortality rates (`esophagus`, `stomach`, `colon`, `liver`, `pancreas`, `lung`, `breast`, `prostate`, `bladder`, `brainCNS`, `thyroid`, `all_leukaemia`, `all_cancer`, `allsolid-NMSC`, `allsolid`, `leukaemia`, `allcause`, `survival`).

Each site-specific data.frame contains variables `age`, `male` and `female`.

For `allcause` mortality data.frame contains age-specific person-year data, `male_py` and `female_py`, in addition to `age`, `male` and `female`.
}
\examples{
 names(Mortality)       # WHO global regions
 names(Mortality[[1]])  # Sites for which baseline mortality rates are available

 # Example 1: All solid cancer mortality rates for Region-1
 Mortality[[1]]$allsolid

 # Example 2: Leukaemia mortality rates for Region-3
 Mortality[[3]]$leukaemia

 # Example 3: A;;ll;-cause mortality rates for Region-5
 Mortality[[5]]$allcause

 # Example 4: plotting lung cancer mortality rates
 plot_refdata( dat=Mortality, outcome="lung", leg_pos=c(0.27,0.95) )

}
\seealso{
\link{Incidence}, \link{plot_refdata}
}
\keyword{datasets}
