% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_refdata.R
\name{plot_riskmodel}
\alias{plot_riskmodel}
\title{Plotting risk models}
\usage{
plot_riskmodel(
  rm,
  doseGy = 0.1,
  maxage = 100,
  ymax = c(1.5, 30),
  leg_pos = "none",
  EAR_PER = 10^4,
  add = c(0, 0),
  sex = "both",
  x_lab = "age (years)",
  y_lab = "excess rate",
  title = ""
)
}
\arguments{
\item{rm}{a list object which contains the risk model information specified in the format described in Details.}

\item{doseGy}{a value that specifies the radiation dose in Gy or Sv (default=0.1).}

\item{maxage}{maximum age in years (default=100)}

\item{ymax}{a value that specifies the maximum value for y-axis.}

\item{leg_pos}{a vector that specifies the legend x-y position ((1,1) for the top-right; default="none" for no-legend).}

\item{EAR_PER}{an integer value for the excess absolute rate denominator (default=10^4).}

\item{add}{a vector of values to be added to female's ERR and EAR values (to avoid completely overlapping sex-specific values when the risk is not modified by sex).}

\item{sex}{a character string which specifies whether the plot contains "both" sexes, "male" or "female" only.}

\item{x_lab}{x-axis label (default="age (years)" ).}

\item{y_lab}{y-axis label (default="excess rate").}

\item{title}{a character string that specifies the title of the plot (default="", i.e., no title).}
}
\value{
a ggplot object
}
\description{
plot_riskmodel() plots the excess relative and absolute rates of cancer incidence or mortality rates based on the risk models derived from the Life Span Study for a given dose. The scenario is currently fixed for ages at exposure of 10, 30 and 50 years. Age specific risks are plotted for both ERR and EAR separately for both male and female.
}
\details{
The risk model can be specified by the list object provided in CanEpiRisk package; for example
LSS_mortality$allsolid$L  for LSS all solid cancer mortality linear model,
LSS_incidence$allsolid$LQ for LSS all solid cancer incidence linear-quadratic model,
LSS_mortality$leukaemia$LQ for LSS leukaemia mortality linear-quadratic model,
LSS_incidence$thyroid$L for LSS thyroid cancer incidence linear model.
}
\examples{
 # The following examples use default data provided in CanEpiRisk package
 # for riskmodels (LSS_mortality and LSS_incidence) derived from Life Span Study
 # and baseline mortality and incidence rates for WHO global regions (Mortality and Incidence)

 # Example 1: LSS all solid cancer mortality risk model
 plot_riskmodel( rm=LSS_mortality$allsolid$L, title="LSS all solid cancer mortality, Linear",  leg_pos=c(0.4, 0.95) )

 # Example 2: LSS Leukaemia incidence risk model
 plot_riskmodel( rm=LSS_incidence$leukaemia$LQ, title="LSS leukaemia incidence", ymax=c(1.5, .3), add=c(0.01,0) )

}
