% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CER.R
\name{CER}
\alias{CER}
\title{Calculate the cumulative excess risk due to radiation exposure}
\usage{
CER(exposure, reference, riskmodel, option, alpha = 0.05)
}
\arguments{
\item{exposure}{exposure scenario (a list object, which contains age(s) at exposure (a single value or a vector), doseGy in Gy or Sv (a single value or a vector),}

\item{reference}{baseline rate and all cause mortality rate in the reference population (a list object, which contains data.frame objects named 'baseline' for baseline rates of the target endpoint and 'mortality' for all cause mortality rates in the reference population)}

\item{riskmodel}{risk model risk model (a list object, which contains two list objects for excess relative risk model (err) and excess absolute risk model (ear), each of which contains a vector of parameter values (para), a matrix of variance covariance matrix (var), and a function to compute the risk given a parameter vector, a dose value, an age at exposure, an attained age and sex.}

\item{option}{option for risk calculation (a list object, which contains maximum age to follow up (an integer value) )}
}
\value{
risk information(vector)
}
\description{
Calculate the cumulative excess risk due to radiation exposure
}
\examples{
   # The following examples use default data provided in the CanEpiRisk package
   # for riskmodel (LSS R14 all solid cancer model),
   #     baseline (all solid cancer mortality rates in Japan 2018) and
   #     mortality  (all cause mortality rates in Japan 2018)

   # Cumulated excess risk for male exposed to 0.1 Gy at age 10
   #  followed up to age 90 with err transfer
   CER( agex=10, doseGy=0.1, sex=1, maxage=90 )

   # Cumulated excess risk for female exposed to 0.01 Gy/year at ages 10-19
   #  followed up to age 100 with 7:3 weights for ERR and EAR transfers
   CER( agex=10:19, doseGy=rep(0.01,10), sex=2, maxage=100, wgt=c(.7,.3) )

}
