% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CER.R
\name{Comp_Exrisk}
\alias{Comp_Exrisk}
\title{Calculating excess risks}
\usage{
Comp_Exrisk(exposure, riskmodel, option, per = 1)
}
\arguments{
\item{exposure}{a list object that specifies the exposure scenario, which contains \code{agex} (a single value or a vector for age(s) at exposure), 'doseGy' (a single value or a vector of dose(s) in Gy), and 'sex' (1 or 2 for male or female).}

\item{riskmodel}{a list object that specifies the risk model, which contains two list objects named \code{err} for excess relative rate model and 'ear' for excess absolute rate model, each of which contains a vector 'para' for model parameter estimates and a function 'f' to compute the excess risk given a parameter vector and exposure information (e.g., dose, age at exposure, sex, attained age).}

\item{option}{a list object that specifies optional settings for risk calculation, which contains an integer value 'maxage' for the maximum age to follow up and a value 'err_wgt' for the weight for risk transfer (1=err, 0=ear).}

\item{per}{an integer value for the risk denominator (default=1).}
}
\value{
information of calculated excess risk (data.frame)
}
\description{
Calculate the excess risk from a risk model under a specified exposure scenario.
}
\examples{
 # The following examples use default data provided in CanEpiRisk package
 # for riskmodels (LSS_mortality and LSS_incidence) derived from Life Span Study
 # and baseline mortality and incidence rates for WHO global regions (Mortality and Incidence).

 # Example 1: allsolid mortality, Region-1, female, 0.1Gy at age 15, followed up to age 100, LSS linear ERR
 exp1 <- list( agex=5, doseGy=0.1, sex=2 )   # exposure scenario
 ref1 <- list( baseline=Mortality[[1]]$allsolid,        # baseline rates
              mortality=Mortality[[1]]$allcause )       # all-cause mortality
 mod1 <- LSS_mortality$allsolid$L                       # risk model
 opt1 <- list( maxage=100, err_wgt=1, n_mcsamp=10000 )  # option
 CER(  exposure=exp1, reference=ref1, riskmodel=mod1, option=opt1 ) * 10000 # cases per 10,000


}
\seealso{
\link{LSS_mortality}, \link{LSS_incidence}
}
