% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Incidence}
\alias{Incidence}
\title{Cancer incidence rates of WHO global regions}
\format{
A list object of the five WHO global regions:
\describe{
  \item{"Aus-NZ Europe Northern America"}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
  \item{"Northern Africa - Western Asia"}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
  \item{"Latin America and Caribbean"}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
  \item{"Asia excl. Western Asia"}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
  \item{"Sub-Saharan Africa"}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
}
}
\source{
IARC
}
\usage{
Incidence
}
\description{
Cancer incidence rates of WHO global regions.
}
\details{
The list object for each region contains the 18 site-specific cancer mortality rates ("esophagus", "stomach", "colon", "liver", "pancreas", "lung", "breast", "prostate", "bladder", "brainCNS", "thyroid", "all_leukaemia", "all_cancer", "allsolid-NMSC", "allsolid", "leukaemia").

Each site-specific data.frame contains variables "age", "male" and "female".
}
\examples{
 # The following examples use default data provided in CanEpiRisk package
 # for riskmodels (LSS_mortality and LSS_incidence) derived from Life Span Study
 # and baseline mortality and incidence rates for WHO global regions (Mortality and Incidence)

 # Example 1: All solid cancer incidence rates for Region-1
 Incidence[[1]]$allsolid

 # Example 2: Leukaemia incidence rates for Region-3
 Incidence[[3]]$leukaemia

 # Example 3: A;;ll;-cause incidence rates for Region-5
 Incidence[[5]]$allcause

 # Example 4: plotting lung cancer incidence rates
 plot_refdata( dat=Incidence, outcome="lung", leg_pos=c(0.27,0.95) )

}
\seealso{
[Mortality], [plot_refdata()]
}
\keyword{datasets}
