% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{LSS_incidence}
\alias{LSS_incidence}
\title{LSS incidence risk models}
\format{
A list object of :
\describe{
\item{\code{allsolid}}{a list object which contains risk model information (see Details)}
\item{\code{leukaemia}}{a list object which contains risk model information (see Details)}
\item{\code{esophagus}}{a list object which contains risk model information (see Details)}
\item{\code{stomach}}{a list object which contains risk model information (see Details)}
\item{\code{colon}}{a list object which contains risk model information (see Details)}
\item{\code{liver}}{a list object which contains risk model information (see Details)}
\item{\code{lung}}{a list object which contains risk model information (see Details)}
\item{\code{prostate}}{a list object which contains risk model information (see Details)}
\item{\code{pancreas}}{a list object which contains risk model information (see Details)}
\item{\code{bladder}}{a list object which contains risk model information (see Details)}
\item{\code{breast}}{a list object which contains risk model information (see Details)}
\item{\code{thyroid}}{a list object which contains risk model information (see Details)}
\item{\code{brainCNS}}{a list object which contains risk model information (see Details)}
}
}
\source{
IARC
}
\usage{
LSS_incidence
}
\description{
LSS incidence risk models for 13 cancer sites.
}
\details{
The list object for each risk model contains the 13 site-specific cancer incidence risk models derived from Life Span Study (\code{leukaemia}, \code{esophagus}, \code{stomach}, \code{colon}, \code{liver}, \code{lung}, \code{prostate}, \code{pancreas}, \code{bladder}, \code{breast}, \code{thyroid}, \code{brainCNS}).

Each site-specific list object contains information for the risk model (a vector of parameter estimates \code{para}, a matrix of variance-covariance matrix of parameter estimates \code{var} and a function to calculate the risk \code{f}).
}
\examples{
 names(LSS_incidence)   # Sites for which LSS incidence risk models are available
 names(LSS_incidence$allsolid)    # Available dose response models
 LSS_incidence$allsolid$L$err     # Linear ERR model for all solid cancer incidence
 LSS_incidence$allsolid$L$ear     # Linear EAR model for all solid cancer incidence

 LSS_incidence$leukaemia$LQ$err   # Linear-quadratic ERR model for leukaemia incidence
 LSS_incidence$lung$L$err         # Linear ERR model for thyroid cancer incidence

 # Plotting LSS all solid cancer incidence risk model
 plot_riskmodel( rm=LSS_incidence$allsolid$L, title="LSS all solid cancer incidence, Linear",  leg_pos=c(0.4, 0.95) )

 # Plotting LSS Leukaemia incidence risk model
 plot_riskmodel( rm=LSS_incidence$leukaemia$LQ, title="LSS leukaemia incidence", ymax=c(1.5, .3), add=c(0.01,0) )

 # Plotting LSS thyroid cancer incidence risk model
 plot_riskmodel( rm=LSS_incidence$thyroid$L, title="LSS thyroid cancer incidence", ymax=c(.5, .3) )

}
\references{
Grant, E.J., A. Brenner, H. Sugiyama et al. Solid Cancer Incidence among the Life Span Study of Atomic Bomb Survivors: 1958-2009. Radiat Res 187(5): 513-537 (2017).

Hsu, W.L., D.L. Preston, M. Soda et al. The incidence of leukemia, lymphoma and multiple myeloma among atomic bomb survivors: 1950-2001. Radiat Res 179(3): 361-382 (2013).

Sakata, R., D.L. Preston, A.V. Brenner et al. Radiation-Related risk of cancers of the upper digestive tract among Japanese atomic bomb survivors. Radiat Res 192(3): 331-344 (2019).

Sadakane, A., B. French, A.V. Brenner et al. Radiation and Risk of Liver, Biliary Tract, and Pancreatic Cancers among Atomic Bomb Survivors in Hiroshima and Nagasaki: 1958-2009. Radiat Res 192(3): 299-310 (2019).

Sugiyama, H., M. Misumi, A. Brenner et al. Radiation risk of incident colorectal cancer by anatomical site among atomic bomb survivors: 1958-2009. Int J Cancer 146(3): 635-645 (2020).

Sugiyama, H., M. Misumi, M. Kishikawa et al. Skin cancer incidence among atomic bomb survivors from 1958 to 1996. Radiat Res 181(5): 531-539 (2014).

Brenner, A.V., D.L. Preston, R. Sakata et al. Incidence of Breast Cancer in the Life Span Study of Atomic Bomb Survivors: 1958-2009. Radiat Res 190(4): 433-444 (2018).

Grant, E.J., M. Yamamura, A.V. Brenner et al. Radiation Risks for the Incidence of Kidney, Bladder and Other Urinary Tract Cancers: 1958-2009. Radiat Res 195(2): 140-148 (2021).

Mabuchi, K., D.L. Preston, A.V. Brenner et al. Risk of Prostate Cancer Incidence among Atomic Bomb Survivors: 1958-2009. Radiat Res 195(1): 66-76 (2021).

Furukawa, K., D. Preston, S. Funamoto et al. Long-term trend of thyroid cancer risk among Japanese atomic-bomb survivors: 60 years after exposure. Int J Cancer 132(5): 1222-1226 (2013).

Brenner, A.V., H. Sugiyama, D.L. Preston et al. Radiation risk of central nervous system tumors in the Life Span Study of atomic bomb survivors, 1958-2009. Eur J Epidemiol 35(6): 591-600 (2020).
}
\seealso{
\link{LSS_mortality}, \link{plot_riskmodel}
}
\keyword{datasets}
