% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{LSS_mortality}
\alias{LSS_mortality}
\title{LSS mortality risk models}
\format{
A list object of :
\describe{
  \item{allsolid}{a list object which contains risk model information (see Details)}
  \item{esophagus}{a list object which contains risk model information (see Details)}
  \item{stomach}{a list object which contains risk model information (see Details)}
  \item{colon}{a list object which contains risk model information (see Details)}
  \item{liver}{a list object which contains risk model information (see Details)}
  \item{lung}{a list object which contains risk model information (see Details)}
  \item{bladder}{a list object which contains risk model information (see Details)}
  \item{breast}{a list object which contains risk model information (see Details)}
  \item{leukaemia}{a list object which contains risk model information (see Details)}
}
}
\source{
IARC
}
\usage{
LSS_mortality
}
\description{
LSS mortality risk models for 9 cancer types.
}
\details{
The list object for each risk model contains the 9 site-specific cancer mortality risk models derived from Life Span Study ("allsolid", "esophagus", "stomach", "colon", "liver", lung", "bladder", "breast", "leukaemia").

Each site-specific data.frame contains information for the risk model (a vector of parameter estimates "para", a matrix of variance-covariance matrix of parameter estimates "var" and a function to calculate the risk "f").
}
\examples{
 names(LSS_mortality)   # Sites for which LSS mortality risk models are available
 names(LSS_mortality$allsolid)    # Available dose response models
 LSS_mortality$allsolid$L$err     # Linear ERR model for all solid cancer motality
 LSS_mortality$allsolid$L$ear     # Linear EAR model for all solid cancer motality

 LSS_mortality$leukaemia$LQ$err   # Linear-quadratic ERR model for leukaemia motality
 LSS_mortality$lung$L$err         # Linear EAR model for lung cancer motality

 # Plotting LSS all solid cancer mortality risk model
 plot_riskmodel( rm=LSS_mortality$allsolid$L, title="LSS all solid cancer mortality, Linear",  leg_pos=c(0.4, 0.95) )
 # Plotting LSS Leukaemia incidence risk model
 plot_riskmodel( rm=LSS_incidence$leukaemia$LQ, title="LSS leukaemia incidence", ymax=c(1.5, .3), add=c(0.01,0) )

}
\references{
Ozasa, K., Y. Shimizu, A. Suyama et al. Studies of the mortality of atomic bomb survivors, Report 14, 1950-2003: an overview of cancer and noncancer diseases. Radiat Res 177(3): 229-243 (2012).
}
\seealso{
[LSS_incidence], [plot_riskmodel()]
}
\keyword{datasets}
