% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Mortality}
\alias{Mortality}
\title{Cancer mortality rates of WHO global regions}
\format{
A list object of the five WHO global regions:
\describe{
\item{\verb{Aus-NZ Europe Northern America}}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
\item{\verb{Northern Africa - Western Asia}}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
\item{\verb{Latin America and Caribbean}}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
\item{\verb{Asia excl. Western Asia}}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
\item{\verb{Sub-Saharan Africa}}{a list object of 18 items, each of which contains a data.frame of cancer rates (see Details)}
}
}
\usage{
Mortality
}
\description{
Cancer mortality rates of WHO global regions.
}
\details{
The list object for each region contains the 18 site-specific cancer mortality rates (\code{esophagus}, \code{stomach}, \code{colon}, \code{liver}, \code{pancreas}, \code{lung}, \code{breast}, \code{prostate}, \code{bladder}, \code{brainCNS}, \code{thyroid}, \code{all_leukaemia}, \code{all_cancer}, \code{allsolid-NMSC}, \code{allsolid}, \code{leukaemia}, \code{allcause}, \code{survival}).

Each site-specific data.frame contains variables \code{age}, \code{male} and \code{female}.

For \code{allcause} mortality data.frame contains age-specific person-year data, \code{male_py} and \code{female_py}, in addition to \code{age}, \code{male} and \code{female}.
}
\examples{
 names(Mortality)       # WHO global regions
 names(Mortality[[1]])  # Sites for which baseline mortality rates are available

 # Example 1: All solid cancer mortality rates for Region-1
 head( Mortality[[1]]$allsolid )
 tail( Mortality[[1]]$allsolid )

 # Example 2: Leukaemia mortality rates for Region-3
 head( Mortality[[3]]$leukaemia )
 tail( Mortality[[3]]$leukaemia )

 # Example 3: A;;ll;-cause mortality rates for Region-5
 head( Mortality[[5]]$allcause )
 tail( Mortality[[5]]$allcause )

 # Example 4: plotting lung cancer mortality rates
 plot_refdata( dat=Mortality, outcome="lung", leg_pos=c(0.27,0.95) )

}
\seealso{
\link{Incidence}, \link{plot_refdata}
}
\keyword{datasets}
