% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CER.R
\name{population_LAR}
\alias{population_LAR}
\title{Calculating population-averaged lifetime attributable risks}
\usage{
population_LAR(
  dsGy,
  reference,
  riskmodel,
  agex = 1:8 * 10 - 5,
  PER = 10000,
  nmc = 10000
)
}
\arguments{
\item{dsGy}{radiation dose in Gy or Sv (a single value).}

\item{reference}{baseline rate, all cause mortality rate and age distribution in the reference population (a list object, which contains data.frame objects named 'baseline' for baseline rates of the target endpoint, 'mortality' for all cause mortality rates and 'agedist' for age distribution in the reference population).}

\item{riskmodel}{risk model risk model (a list object, which contains two list objects for excess relative risk model (err) and excess absolute risk model (ear), each of which contains a vector of parameter values (para), a matrix of variance covariance matrix (var), and a function to compute the risk given a parameter vector, a dose value, an age at exposure, an attained age and sex.}

\item{agex}{a vector of ages at exposure, which represent age categories (default values: 5, 15, 25, ..., 75 to represent age categories 0-10, 10-20, ..., 70-80) option for risk calculation (a list object, which contains maximum age to follow up (an integer value)}

\item{PER}{an integer (default value: 10000 to show the estimated risk as cases per 10000)}

\item{mmc}{an integer for the Monte Carlo sample size (default: 10000)}
}
\value{
estimated risk information (list)
}
\description{
Calculate the population-averaged lifetime attributable risk due to radiation exposure.
}
\examples{
 # The following examples use default data provided in CanEpiRisk package
 # for riskmodels (LSS_mortality and LSS_incidence) derived from Life Span Study
 # and baseline mortality and incidence rates for WHO global regions (Mortality and Incidence).

 # Example 1: allsolid mortality, Region-1, exposed to 0.1 Gy, followed up to age 100, LSS linear ERR
 ref1 <- list(  baseline=Mortality[[1]]$allsolid,     # baseline rates
               mortality=Mortality[[1]]$allcause,     # allcause mortality
                 agedist=agedist_rgn[[1]] )           # age distribution
 mod1 <- LSS_mortality$allsolid$L                     # risk model
 population_LAR( dsGy=0.1, reference=ref1, riskmodel=mod1 )    # CER cases per 10,000

 # Example 2: leukaemia incidence, Region-4, exposed to 0.1 Gy, followed up to age 100, LSS LQ ERR
 ref2 <- list(  baseline=Incidence[[4]]$leukaemia,    # baseline rates
               mortality=Mortality[[4]]$allcause,     # all-cause mortality
                 agedist=agedist_rgn[[4]] )           # age distribution
 mod2 <- LSS_incidence$leukaemia$LQ                   # risk model
 population_LAR( dsGy=0.1, reference=ref2, riskmodel=mod2 )    # CER cases per 10,000

}
\seealso{
\link{CER}, \link{population_YLL}
}
